refactor(DK): 重构蓝牙数据传输格式并优化相关功能
- 修改 BT_Init() 中的波特率设置，从 115200 降至 9600
- 重写 BT_SendDataPacket() 函数，使用浮点数表示湿度和温度
- 更新 CountSensor 相关文件，将计数值改为 8 位无符号整数
- 调整主程序逻辑，以适应新的数据格式和传输方式

+++ DK/CountSensor.c
@@ -1,7 +1,7 @@
 #include "stm32f10x.h" // Device header
 #include "dk_C8T6.h"
 
-uint16_t CountSensor_Count; // 全局变量，用于计数
+uint8_t CountSensor_Count;
 
 /**
  * 函    数：计数传感器初始化
@@ -53,7 +53,7 @@ void CountSensor_Init(void)
  * 参    数：无
  * 返 回 值：计数值，范围：0~65535
  */
-uint16_t CountSensor_Get(void)
+uint8_t CountSensor_Get(void)
 {
     return CountSensor_Count;
 }
