+++ DK/CountSensor.h
@@ -1,9 +1,9 @@
 #ifndef __COUNT_SENSOR_H
 #define __COUNT_SENSOR_H
 
-extern uint16_t CountSensor_Count; // 计数值
+extern uint8_t CountSensor_Count; // 计数值
 
 void CountSensor_Init(void);
-uint16_t CountSensor_Get(void);
+uint8_t CountSensor_Get(void);
 
 #endif // __COUNT_SENSOR_H
